Add Payee broken

+++ src/main/java/datamodel/UserDataModel.java
@@ -7,6 +7,7 @@ import model.Transaction;
 import model.User;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
 
@@ -29,7 +30,7 @@ public interface UserDataModel {
 
     void makeTransaction(BankAccount accountToUse, BigDecimal amount, Beneficiary beneficiary);
 
-    void addBeneficiary(long beneficiaryAccount, long userAccount);
+    void addBeneficiary(long beneficiaryAccount, BigInteger userAccount);
 
     Observable<Boolean> getAddBeneficiarySuccessStream();
